+++ src/main/java/pl/ergohestia/ehj1/ivesta/model/DriverHeaders.java
@@ -1,3 +1,3 @@
-package pl.ergohestia.ehj1.ivesta.utils;
+package pl.ergohestia.ehj1.ivesta.model;
 
 public enum DriverHeaders {imie, nazwisko, adres, telefon, licencja, liczba_kursow, liczba_kilometrow}
